Some random changes:
* Move the identity image styling into the stylesheet. It looked a bit large.
* Show the current user's username instead of having everyone be Copper alts.
* Show a front page instead of the feed if the user isn't logged in. The
  content is very temporary and we should probably put something that looks
  like the current "elevator pitch page" there instead.

Yes, this breaks everything. As in you will never show as being logged in
even though you are using the right username/password, etc. In theory these
changes should work fine, but there's a bug in the way login is done.

When we call

    return HTTPMoved(request.url("home"))

in `POST_login()`, it's treated like a bad request (401 or similar). One of
the side effects of this is that the callbacks that are normally run when a
request completes are skipped. This includes the callback that would persist
the user information in Redis, so you never actually log in.

I tried forcing the session to save in `POST_login()` but Distill doesn't
like that. The rendering pipeline chokes when it gets the `HTTPMoved` because
it's an exception and doesn't behave the same way as a normal request.